cleanup_all.py
"""
Cleanup Script - Clear all deal rooms and delete unwanted files
"""
import os
import asyncio
from telegram import Bot
from config import BOT_TOKEN, ROOM_POOL

async def cleanup_rooms():
    """Clear all messages from deal rooms"""
    bot = Bot(token=BOT_TOKEN)
    
    print("=" * 60)
    print("CLEANING UP DEAL ROOMS")
    print("=" * 60)
    
    for room_id in ROOM_POOL:
        try:
            print(f"\nCleaning room: {room_id}")
            
            # Get chat info
            chat = await bot.get_chat(room_id)
            print(f"  Room name: {chat.title}")
            
            # Delete all messages (Telegram doesn't allow bulk delete, so we skip this)
            # Instead, we'll just send a cleanup message
            await bot.send_message(
                room_id,
                "🧹 *Room Cleaned*\n\nAll previous deals cleared.\nRoom is now available for new deals.",
                parse_mode="Markdown"
            )
            
            print(f"  ✅ Cleanup message sent")
            
        except Exception as e:
            print(f"  ❌ Error: {e}")
    
    print("\n" + "=" * 60)
    print("CLEANUP COMPLETE")
    print("=" * 60)

def delete_unwanted_files():
    """Delete test files and temporary files"""
    print("\n" + "=" * 60)
    print("DELETING UNWANTED FILES")
    print("=" * 60)
    
    # List of files to delete
    unwanted_files = [
        'test_api_now.py',
        'test_api_simple.py',
        'test_bscscan_api.py',
        'test_without_key.py',
        'test_real_transaction.py',
        'test_monitor.py',
        'test_bot.py',
        'test_web3_monitor.py',
        'auto_test.py',
        'quick_test.py',
        'run_complete_test.bat',
        'test_output.txt',
        'test_results.txt',
        'API_KEY_ISSUE_SOLUTION.md',
        'GET_BSCSCAN_API_KEY.md',
        'SIMPLE_API_KEY_GUIDE.md',
        'FINAL_API_SOLUTION.md',
        'AUTO_DETECTION_WORKING.md',
        'bot.py',  # Old version
        'bot_v2.py',  # Old version
        'bot_complete.py',  # Old version
        'blockchain_monitor.py',  # Old API-based version
        'clear_and_check.py',
        'update_wallet.py',
        'UPDATE_WALLET.txt',
        'FINAL_SUMMARY.md',
        'IMPLEMENTATION_SUMMARY.md',
        'TODO.md',
        'INSTALLATION.md',
        'START_BOT.md',
        'FINAL_SETUP_GUIDE.md',
        'setup_guide.txt'
    ]
    
    deleted_count = 0
    for filename in unwanted_files:
        filepath = os.path.join(os.path.dirname(__file__), filename)
        if os.path.exists(filepath):
            try:
                os.remove(filepath)
                print(f"  ✅ Deleted: {filename}")
                deleted_count += 1
            except Exception as e:
                print(f"  ❌ Failed to delete {filename}: {e}")
        else:
            print(f"  ⏭️  Skipped: {filename} (not found)")
    
    print(f"\n  Total deleted: {deleted_count} files")
    print("=" * 60)

async def main():
    print("\n🧹 COMPLETE CLEANUP SCRIPT")
    print("This will:")
    print("1. Clear all deal room messages")
    print("2. Delete unwanted test files")
    print("\n")
    
    # Cleanup rooms
    await cleanup_rooms()
    
    # Delete files
    delete_unwanted_files()
    
    print("\n✅ ALL CLEANUP COMPLETE!")
    print("\nYour bot is now clean and ready for production!")
    print("\nRemaining files:")
    print("  - bot_main.py (main bot)")
    print("  - blockchain_monitor_web3.py (Web3 monitor)")
    print("  - transaction_handler.py")
    print("  - auth_system.py")
    print("  - room_manager.py")
    print("  - config.py")
    print("  - .env")
    print("  - requirements.txt")
    print("  - README.md")
    print("  - WEB3_SOLUTION_FINAL.md")
    print("  - HOW_TO_RUN.md")
    print("  - QUICK_REFERENCE.md")

if __name__ == '__main__':
    asyncio.run(main())
