requirements.txt
python-telegram-bot==21.10
python-dotenv==1.0.0
web3==6.11.0
eth-account==0.10.0
requests==2.31.0
